+++ packages/web-components/keen-slider-auto-switch/src/components/keen-slider-auto-switch.ts
@@ -1,17 +1,14 @@
-import type { CorrectComponentType } from "@brytdesigns/web-component-utils";
+import type { CorrectComponentType } from "@brytdesigns/web-component-core/utils";
 
 import {
   createEffect,
   createSignal,
-  createMemo,
   on,
   onCleanup,
   batch,
+  untrack,
 } from "solid-js";
-import {
-  addPlugin,
-  type KeenSliderInstance,
-} from "@brytdesigns/web-component-keen-slider";
+import { withKeenSliderElementContext } from "@brytdesigns/web-component-keen-slider";
 
 type Props = {
   target: string;
@@ -26,127 +23,124 @@ export const Component: CorrectComponentType<Props> = (props, { element }) => {
       `${Name}: Needs a proper target in order to properly extend a keen slider.`,
     );
 
-  const target = createMemo(() => {
-    let targetEl: HTMLElement | null = element;
-    if (props.target) targetEl = document.querySelector(props.target);
-    if (!targetEl)
-      return console.warn(
-        `${Name}: Could not find the target element. Make sure it exists and is a keen-slider element.`,
+  withKeenSliderElementContext(
+    () => {
+      const selector = props.target;
+      return () => {
+        let targetEl: Element | null = element;
+        if (props.target) targetEl = document.querySelector(selector);
+        if (targetEl?.tagName !== "KEEN-SLIDER")
+          targetEl = targetEl!.querySelector("keen-slider");
+        return targetEl!;
+      };
+    },
+    () => props.duration,
+    (context, duration) => {
+      const [sliderRef] = context;
+      const [_, setTimeoutVal] = createSignal<number | null>(null);
+      const [isHovering, setIsHovering] = createSignal(false);
+      const [isPaused, setIsPaused] = createSignal(false);
+
+      function clearCurrentTimeout() {
+        batch(() => {
+          setIsPaused(true);
+          setTimeoutVal((val) => {
+            if (val) clearTimeout(val);
+            return null;
+          });
+        });
+      }
+
+      function startNextTimeout(cb: any, duration: number) {
+        if (untrack(isHovering)) return;
+        batch(() => {
+          setIsPaused(false);
+          setTimeoutVal((val) => {
+            if (val) clearTimeout(val);
+            return setTimeout(cb, duration) as any as number;
+          });
+        });
+      }
+
+      createEffect(
+        on(
+          () => ({ isPaused: isPaused(), slider: sliderRef() }),
+          ({ isPaused, slider }) => {
+            if (!slider) return;
+            slider.container.dispatchEvent(
+              new CustomEvent("autoswitch:update", {
+                bubbles: true,
+                detail: { isPaused },
+              }),
+            );
+          },
+        ),
       );
-    if (targetEl.tagName !== "KEEN-SLIDER")
-      targetEl = targetEl.querySelector("keen-slider");
-    if (!targetEl)
-      return console.warn(
-        `${Name}: Could not find the target element. Make sure it exists and is a keen-slider element.`,
+
+      createEffect(
+        on(
+          () => ({ slider: sliderRef() }),
+          ({ slider }) => {
+            if (!slider) return;
+            const controller = new AbortController();
+
+            function nextTimeout() {
+              if (!slider) return;
+              if (slider.options?.disabled) return;
+              return startNextTimeout(() => {
+                if (slider.slides.length <= 1) return;
+                slider.next();
+              }, duration);
+            }
+
+            (slider as any).pauseAutoplay = clearCurrentTimeout;
+            (slider as any).startAutoplay = nextTimeout;
+
+            slider.container.addEventListener(
+              "mouseenter",
+              () => {
+                batch(() => {
+                  setIsHovering(true);
+                  clearCurrentTimeout();
+                });
+              },
+              {
+                signal: controller.signal,
+              },
+            );
+            slider.container.addEventListener(
+              "mouseleave",
+              () => {
+                batch(() => {
+                  setIsHovering(false);
+                  startNextTimeout(nextTimeout, duration);
+                });
+              },
+              {
+                signal: controller.signal,
+              },
+            );
+
+            slider.on("dragStarted", clearCurrentTimeout);
+            slider.on("slideChanged", clearCurrentTimeout);
+            slider.on("animationEnded", nextTimeout);
+            slider.on("updated", nextTimeout);
+            slider.on("destroyed", clearCurrentTimeout, true);
+
+            nextTimeout();
+
+            return onCleanup(() => {
+              controller.abort();
+
+              slider.on("dragStarted", clearCurrentTimeout, true);
+              slider.on("slideChanged", clearCurrentTimeout, true);
+              slider.on("animationEnded", nextTimeout, true);
+              slider.on("updated", nextTimeout, true);
+              slider.on("destroyed", clearCurrentTimeout, true);
+            });
+          },
+        ),
       );
-    return targetEl;
-  });
-
-  const [sliderRef, setSliderRef] = createSignal<KeenSliderInstance | null>(
-    null,
+    },
   );
-  const [_, setTimeoutVal] = createSignal<number | null>(null);
-  const [isPaused, setIsPaused] = createSignal(false);
-
-  function clearCurrentTimeout() {
-    batch(() => {
-      setIsPaused(true);
-      setTimeoutVal((val) => {
-        if (val) clearTimeout(val);
-        return null;
-      });
-    });
-  }
-
-  function startNextTimeout(cb: any, duration: number) {
-    batch(() => {
-      setIsPaused(false);
-      setTimeoutVal((val) => {
-        if (val) clearTimeout(val);
-        return setTimeout(cb, duration) as any as number;
-      });
-    });
-  }
-
-  createEffect(
-    on(
-      () => ({ isPaused: isPaused(), slider: sliderRef() }),
-      ({ isPaused, slider }) => {
-        if (!slider) return;
-        slider.container.dispatchEvent(
-          new CustomEvent("autoswitch:update", {
-            bubbles: true,
-            detail: { isPaused },
-          }),
-        );
-      },
-    ),
-  );
-
-  createEffect(
-    on(
-      () => ({ slider: sliderRef(), duration: props.duration }),
-      ({ slider, duration }) => {
-        if (!slider) return;
-        const controller = new AbortController();
-
-        function nextTimeout() {
-          if (!slider) return;
-          if (slider.options?.disabled) return;
-          return startNextTimeout(() => {
-            if (slider.slides.length <= 1) return;
-            slider.next();
-          }, duration);
-        }
-
-        (slider as any).pauseAutoplay = clearCurrentTimeout;
-        (slider as any).startAutoplay = nextTimeout;
-
-        slider.container.addEventListener("mouseenter", clearCurrentTimeout, {
-          signal: controller.signal,
-        });
-        slider.container.addEventListener("mouseleave", nextTimeout, {
-          signal: controller.signal,
-        });
-
-        slider.on("dragStarted", clearCurrentTimeout);
-        slider.on("slideChanged", clearCurrentTimeout);
-        slider.on("animationEnded", nextTimeout);
-        slider.on("updated", nextTimeout);
-
-        nextTimeout();
-
-        return onCleanup(() => {
-          controller.abort();
-
-          slider.on("dragStarted", clearCurrentTimeout, true);
-          slider.on("slideChanged", clearCurrentTimeout, true);
-          slider.on("animationEnded", nextTimeout, true);
-          slider.on("updated", nextTimeout, true);
-        });
-      },
-    ),
-  );
-
-  createEffect(() => {
-    const t = target();
-
-    if (!t) return;
-
-    const plugin = (slider: KeenSliderInstance | null) => {
-      setSliderRef(slider);
-    };
-
-    const controller = new AbortController();
-
-    addPlugin({
-      target: t,
-      plugin,
-      controller,
-    });
-
-    return onCleanup(() => {
-      controller.abort();
-    });
-  });
 };