+++ packages/web-components/keen-slider-navigation-dots/src/components/index.ts
@@ -1,1 +1,1 @@
-export * from "./keen-slider-navigation-dots.js";
+export * as KeenSliderNavigationDots from "./keen-slider-navigation-dots.js";