+++ packages/web-components/keen-slider-scrollbar/src/components/keen-slider-scrollbar.ts
@@ -1,4 +1,4 @@
-import type { CorrectComponentType } from "@brytdesigns/web-component-utils";
+import type { CorrectComponentType } from "@brytdesigns/web-component-core/utils";
 
 import html from "solid-js/html";
 import {
@@ -7,12 +7,13 @@ import {
   createMemo,
   on,
   onCleanup,
+  batch,
   Show,
   For,
   type Accessor,
 } from "solid-js";
 import {
-  addPlugin,
+  withKeenSliderElementContext,
   type KeenSliderInstance,
 } from "@brytdesigns/web-component-keen-slider";
 import { Draggable } from "@neodrag/vanilla";
@@ -27,35 +28,19 @@ import {
   srOnly,
 } from "../utils.js";
 
-type KeenSliderScrollbarProps = {
+type Props = {
   target: string;
   isHidden: boolean;
   isVertical: boolean;
 };
 
-export const KeenSliderScrollbar: CorrectComponentType<
-  KeenSliderScrollbarProps
-> = (props, { element }) => {
+export const Name = `keen-slider-scrollbar`;
+
+export const Component: CorrectComponentType<Props> = (props, { element }) => {
   if (!props.target)
     return console.warn(
-      "keen-slider-scrollbar: Needs a proper target in order to properly extend a keen slider.",
+      `${Name}: Needs a proper target in order to properly extend a keen slider.`,
     );
-  const target = createMemo(() => {
-    let targetEl: HTMLElement | null = element;
-    if (props.target) targetEl = document.querySelector(props.target);
-    if (!targetEl)
-      return console.warn(
-        "keen-slider-scrollbar: Could not find the target element. Make sure it exists and is a keen-slider element.",
-      );
-    if (targetEl.tagName !== "KEEN-SLIDER")
-      targetEl = targetEl.querySelector("keen-slider");
-    if (!targetEl)
-      return console.warn(
-        "keen-slider-scrollbar: Could not find the target element. Make sure it exists and is a keen-slider element.",
-      );
-
-    return targetEl;
-  });
 
   const [slider, setSlider] = createSignal<KeenSliderInstance>();
 
@@ -244,30 +229,35 @@ export const KeenSliderScrollbar: CorrectComponentType<
     }),
   );
 
-  createEffect(() => {
-    const t = target();
-    if (!t) return;
+  withKeenSliderElementContext(
+    () => {
+      const selector = props.target;
+      return () => {
+        let targetEl: Element | null = element;
+        if (props.target) targetEl = document.querySelector(selector);
+        if (targetEl?.tagName !== "KEEN-SLIDER")
+          targetEl = targetEl!.querySelector("keen-slider");
+        return targetEl!;
+      };
+    },
+    () => null,
+    ([_, { addPlugin }]) => {
+      addPlugin((slider) => {
+        setSlider(slider);
 
-    const controller = new AbortController();
+        slider.on("created", function (slider) {
+          setMaxIdx(slider.track.details.maxIdx);
+        });
 
-    function plugin(slider: KeenSliderInstance) {
-      setSlider(slider);
-      slider.on("created", function (slider) {
-        setMaxIdx(slider.track.details.maxIdx);
+        slider.on("destroyed", () =>
+          batch(() => {
+            setMaxIdx(0);
+            setSlider();
+          }),
+        );
       });
-    }
-
-    addPlugin({
-      target: t,
-      plugin,
-      controller,
-    });
-
-    return onCleanup(() => {
-      controller.abort();
-      setSlider(undefined);
-    });
-  });
+    },
+  );
 
   function animateScrollbarPosition(p: { x: number; y: number }) {
     animate({
@@ -289,7 +279,7 @@ export const KeenSliderScrollbar: CorrectComponentType<
   function updateSliderPosition(bp: Element, slider: KeenSliderInstance) {
     const idx = constrain(
       parseInt(bp.getAttribute("data-index") || "0") *
-      ((slider?.options?.slides as any)?.perScroll || 1),
+        ((slider?.options?.slides as any)?.perScroll || 1),
       0,
       slider.track.details.maxIdx,
     );