+++ packages/web-components/keen-slider/src/components/keen-slider.ts
@@ -1,4 +1,4 @@
-import type { CorrectComponentType } from "@brytdesigns/web-component-utils";
+import type { CorrectComponentType } from "@brytdesigns/web-component-core/utils";
 
 import { observable } from "solid-js";
 import _KeenSlider, {
@@ -11,28 +11,27 @@ import {
   useKeenSliderContext,
 } from "../hooks/index.js";
 
-type KeenSliderContextProps = {
+type Props = {
   config?: KeenSliderOptions;
   centerSlides: boolean;
   refreshOnChildrenChange: boolean;
 };
 
-export const KeenSlider: CorrectComponentType<KeenSliderContextProps> = (
-  props,
-  { element }
-) => {
+export const Name = `keen-slider`;
+
+export const Component: CorrectComponentType<Props> = (props, { element }) => {
   const context = provideKeenSliderContext(props, element);
   const [slider, methods] = useKeenSliderContext(context);
 
   const observer = observable(slider);
 
   Object.assign(element, {
-    subscribe(callback = (slider: KeenSliderInstance) => null) {
+    subscribe(callback = (_: KeenSliderInstance) => null) {
       function listen(slider: KeenSliderInstance) {
         const cleanup = callback(slider);
         slider.on(
           "destroyed",
-          typeof cleanup === "function" ? cleanup : () => null
+          typeof cleanup === "function" ? cleanup : () => null,
         );
       }
       return observer.subscribe(listen).unsubscribe;