+++ examples/testing/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @example/testing
 
+## 0.1.2
+
+### Patch Changes
+
+- Minor bug fix with context cleanup and upgraded keen slider components to new core library
+
 ## 0.1.1
 
 ### Patch Changes