+++ packages/web-components/keen-slider-scrollbar/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @brytdesigns/web-component-keen-slider-scrollbar
 
+## 1.1.0
+
+### Minor Changes
+
+- Minor bug fix with context cleanup and upgraded keen slider components to new core library
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/web-component-keen-slider@1.1.0
+  - @brytdesigns/web-component-core@1.1.2
+
 ## 1.0.20
 
 ### Patch Changes