+++ shopify/packages/product-info-components/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brytdesigns/shopify-product-info-components
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/web-component-core@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes